Latest version with news controllers added

+++ Controllers/EmployeesController.php
@@ -3,11 +3,12 @@
 namespace Employees\Controllers;
 
 
+
 use Employees\Models\Binding\Emp\EmpBindingModel;
+use Employees\Services\AuthenticationServiceInterface;
 use Employees\Services\CreatingQueryServiceInterface;
 use Employees\Services\EmployeesServiceInterface;
 use Employees\Services\EncryptionServiceInterface;
-use Employees\Services\CreatingQuerySevice;
 
 class EmployeesController
 {
@@ -15,12 +16,17 @@ class EmployeesController
     private $employeeService;
     private $encryptionService;
     private $createQuery;
+    private $authenticationService;
 
-    public function __construct(EmployeesServiceInterface $employeesService, EncryptionServiceInterface $encryptionService, CreatingQueryServiceInterface $createQuery)
+    public function __construct(EmployeesServiceInterface $employeesService,
+                                EncryptionServiceInterface $encryptionService,
+                                CreatingQueryServiceInterface $createQuery,
+                                AuthenticationServiceInterface $authenticationService)
     {
         $this->employeeService = $employeesService;
         $this->encryptionService = $encryptionService;
         $this->createQuery = $createQuery;
+        $this->authenticationService = $authenticationService;
     }
 
     public function option()
@@ -30,26 +36,32 @@ class EmployeesController
 
     public function list($active = null)
     {
-        if ($active == null) {
 
-            print_r(json_encode(array("employee" => $this->employeeService->getListStatus("yes"))));
+        //if ($this->authenticationService->isTokenCorrect()) {
 
-        } else {
+            if ($active == null) {
 
-            print_r(json_encode(array("employee" => $this->employeeService->getListStatus($active))));
+                print_r(json_encode(array("employee" => $this->employeeService->getListStatus("yes"))));
 
-        }
+            } else {
+
+                print_r(json_encode(array("employee" => $this->employeeService->getEmp($active))));
+
+            }
+
+        //}
 
     }
 
     public function removeEmployee($id) {
+         if ($this->authenticationService->isTokenCorrect()) {
 
-        if ($this->employeeService->removeEmp($id)) {
-            print_r("true");
-        } else {
-            print_r("false");
-        }
-
+            if ($this->employeeService->removeEmp($id)) {
+                print_r("true");
+            } else {
+                print_r("false");
+            }
+         }
     }
 
 
@@ -58,6 +70,9 @@ class EmployeesController
 
 //        var_dump("TEST");
 //        exit;
+        if ($this->authenticationService->isTokenCorrect()) {
+
+
         $md5string = $this->encryptionService->md5generator($employeeBindingModel->getFirstName().
             $employeeBindingModel->getLastName().
             $employeeBindingModel->getBirthday());
@@ -75,6 +90,8 @@ class EmployeesController
         } else {
             print_r("false");
         }
+
+        }
     }
 
     public function getemployee($id) {
@@ -83,18 +100,23 @@ class EmployeesController
 
     }
 
-    public function updateemployee($theid,EmpBindingModel $empBindingModel) {
+    public function updateemployee($theid, EmpBindingModel $empBindingModel)
+    {
 
-        $empBindingModel->setId($theid);
+        //if ($this->authenticationService->isTokenCorrect()) {
 
-        $this->createQuery->setQueryUpdateEmp($empBindingModel);
+            $empBindingModel->setId($theid);
 
-//        if ($this->employeeService->updEmp($empBindingModel)) {
-        if ($this->employeeService->updEmp($this->createQuery->getQuery(), $this->createQuery->getValues())) {
-            print_r(json_encode(array("employees" => $this->employeeService->getEmp($empBindingModel->getId()))));
-        } else {
-            print_r("false");
-        }
+            //        if ($this->employeeService->updEmp($empBindingModel)) {
+            if ($this->employeeService->updEmp($empBindingModel)) {
+                print_r(json_encode(array("employees" => $this->employeeService->getEmp($empBindingModel->getId()))));
+            } else {
+                print_r("false");
+            }
+
+//           } else {
+//            http_response_code("404");
+//        }
     }
 
 }
